feat: remove redundant files and move bottom border extension to modifier extensions file
